Map refactor; split into smaller components
