+++ src/components/SharedComponents/Map/Map.js
@@ -1,12 +1,9 @@
 // @flow
 
-import AsyncStorage from "@react-native-async-storage/async-storage";
 import { useNavigation } from "@react-navigation/native";
 import classnames from "classnames";
-import { INatIconButton } from "components/SharedComponents";
 import LocationPermissionGate from "components/SharedComponents/LocationPermissionGate";
 import { View } from "components/styledComponents";
-import LocationIndicator from "images/svg/location_indicator.svg";
 import type { Node } from "react";
 import React, {
   useCallback,
@@ -15,76 +12,24 @@ import React, {
   useRef,
   useState
 } from "react";
-import MapView, {
-  Circle,
-  Marker,
-  Polygon,
-  UrlTile
-} from "react-native-maps";
-import { useTheme } from "react-native-paper";
-import createUTFPosition from "sharedHelpers/createUTFPosition";
-import getDataForPixel from "sharedHelpers/fetchUTFGridData";
+import MapView from "react-native-maps";
 import { useDeviceOrientation } from "sharedHooks";
-import useTranslation from "sharedHooks/useTranslation";
-import { getShadowForColor } from "styles/global";
-import colors from "styles/tailwindColors";
 
-const DROP_SHADOW = getShadowForColor( colors.darkGray );
+import CurrentLocationButton from "./CurrentLocationButton";
+import {
+  calculateZoom,
+  fetchObservationUUID,
+  metersToLatitudeDelta,
+  OBSCURATION_CELL_SIZE,
+  obscurationCellForLatLng,
+  zoomToDeltas
+} from "./helpers/mapHelpers";
+import LocationIndicator from "./LocationIndicator";
+import ObscuredLocationIndicator from "./ObscuredLocationIndicator";
+import ObsUrlTile from "./ObsUrlTile";
+import SwitchMapTypeButton from "./SwitchMapTypeButton";
 
-function calculateZoom( width, delta ) {
-  return Math.round(
-    Math.log2( 360 * ( width / 256 / delta ) ) + 1
-  );
-}
-
-// Kind of the inverse of calculateZoom. Probably not actually accurate for
-// longitude, but works for our purposes
-function zoomToDeltas( zoom, screenWidth, screenHeight ) {
-  const longitudeDelta = screenWidth / 256 / ( 2 ** zoom / 360 );
-  const latitudeDelta = screenHeight / 256 / ( 2 ** zoom / 360 );
-  return [latitudeDelta, longitudeDelta];
-}
-
-const POINT_TILES_ENDPOINT = "https://tiles.inaturalist.org/v1/points";
-const API_ENDPOINT = "https://api.inaturalist.org/v2";
-const OBSCURATION_CELL_SIZE = 0.2;
 const NEARBY_DIM_M = 50_000;
-
-// Adapted from
-// https://github.com/inaturalist/inaturalist/blob/main/app/assets/javascripts/inaturalist/map3.js.erb#L1500
-function obscurationCellForLatLng( lat, lng ) {
-  const coords = [lat, lng];
-  const firstCorner = [
-    coords[0] - ( coords[0] % OBSCURATION_CELL_SIZE ),
-    coords[1] - ( coords[1] % OBSCURATION_CELL_SIZE )
-  ];
-  const secondCorner = [firstCorner[0], firstCorner[1]];
-  coords.forEach( ( value, index ) => {
-    if ( value < secondCorner[index] ) {
-      secondCorner[index] -= OBSCURATION_CELL_SIZE;
-    } else {
-      secondCorner[index] += OBSCURATION_CELL_SIZE;
-    }
-  } );
-  return {
-    minLat: Math.min( firstCorner[0], secondCorner[0] ),
-    minLng: Math.min( firstCorner[1], secondCorner[1] ),
-    maxLat: Math.max( firstCorner[0], secondCorner[0] ),
-    maxLng: Math.max( firstCorner[1], secondCorner[1] )
-  };
-}
-
-// Adapted from iNat Android LocationChooserActivity.java computeOffset function
-const EARTH_RADIUS = 6371000; // Earth radius in meters
-export function metersToLatitudeDelta( meters: number, latitude: number ): number {
-  // Calculate latitude delta in radians
-  const latitudeDeltaRadians
-    = meters / ( EARTH_RADIUS * Math.cos( ( latitude * Math.PI ) / 180 ) );
-
-  // Convert latitude delta to degrees
-  const latitudeDelta = ( latitudeDeltaRadians * 180 ) / Math.PI;
-  return latitudeDelta;
-}
 
 type Props = {
   // $FlowIgnore
@@ -181,11 +126,9 @@ const Map = ( {
       : 5
   );
   const navigation = useNavigation( );
-  const theme = useTheme( );
   const [permissionRequested, setPermissionRequested] = useState( permissionRequestedProp );
   const [showsUserLocation, setShowsUserLocation] = useState( false );
   const [userLocation, setUserLocation] = useState( null );
-  const { t } = useTranslation( );
   const mapViewRef = useRef( mapViewRefProp );
   const [currentMapType, setCurrentMapType] = useState( mapType || "standard" );
 
@@ -227,16 +170,6 @@ const Map = ( {
       setZoomToNearbyRequested( true );
     }
   }, [startAtNearby, zoomToNearbyRequested] );
-
-  useEffect( () => {
-    AsyncStorage.getItem( "mapType" ).then( value => {
-      if ( value && !mapType ) {
-        // Load last saved map type (unless explicitly overridden by the parent
-        // of the Map component)
-        setCurrentMapType( value );
-      }
-    } );
-  }, [mapType, setCurrentMapType] );
 
   useEffect( ( ) => {
     if ( userLocation && zoomToUserLocationRequested && mapViewRef?.current ) {
@@ -342,11 +275,6 @@ const Map = ( {
     return newTileParams;
   }, [tileMapParams] );
 
-  const changeMapType = async newMapType => {
-    setCurrentMapType( newMapType );
-    await AsyncStorage.setItem( "mapType", newMapType );
-  };
-
   const obscurationCell = obscurationCellForLatLng( obsLatitude, obsLongitude );
   if ( obscured ) {
     initialRegion = {
@@ -357,83 +285,12 @@ const Map = ( {
     };
   }
 
-  const queryString = Object.keys( params ).map( key => `${key}=${params[key]}` ).join( "&" );
-
-  const showPointTiles = currentZoom > 13;
-  // We want green points and (default) orange grid
-  const tileUrlTemplate = showPointTiles
-    ? `${API_ENDPOINT}/points/{z}/{x}/{y}.png?${queryString}&color=%2374ac00`
-    : `${API_ENDPOINT}/grid/{z}/{x}/{y}.png?${queryString}`;
-
   const onMapPressForObsLyr = useCallback( async latLng => {
-    const UTFPosition = createUTFPosition( currentZoom, latLng.latitude, latLng.longitude );
-    const {
-      mTilePositionX,
-      mTilePositionY,
-      mPixelPositionX,
-      mPixelPositionY
-    } = UTFPosition;
-    const tilesParams = {
-      ...params,
-      style: "geotilegrid"
-    };
-    const gridQuery = Object.keys( tilesParams )
-      .map( key => `${key}=${tilesParams[key]}` ).join( "&" );
-
-    const gridUrl = `${POINT_TILES_ENDPOINT}/${currentZoom}/${mTilePositionX}/${mTilePositionY}`
-      + ".grid.json";
-    const gridUrlTemplate = `${gridUrl}?${gridQuery}`;
-
-    const options = {
-      method: "GET",
-      headers: {
-        Accept: "application/json"
-      }
-    };
-
-    const response = await fetch( gridUrlTemplate, options );
-    const json = await response.json( );
-
-    const observation = getDataForPixel( mPixelPositionX, mPixelPositionY, json );
-    const uuid = observation?.uuid;
-
+    const uuid = fetchObservationUUID( currentZoom, latLng, params );
     if ( uuid ) {
       navigation.navigate( "ObsDetails", { uuid } );
     }
   }, [params, currentZoom, navigation] );
-
-  const locationIndicator = () => (
-    // $FlowIgnore
-    <LocationIndicator
-      testID="Map.LocationIndicator"
-      width={25}
-      height={32}
-    />
-  );
-
-  // Not clear why but nesting <UrlTile> directly under <MapView> seems to
-  // cause it not to update in Android when you change the URL
-  const ObsUrlTile = useCallback( ( ) => {
-    if ( !tileUrlTemplate ) return <View />;
-    if ( !withPressableObsTiles && !withObsTiles ) return <View />;
-    return (
-      <UrlTile
-        testID="Map.UrlTile"
-        tileSize={512}
-        urlTemplate={tileUrlTemplate}
-        opacity={
-          showPointTiles
-            ? 1
-            : 0.7
-        }
-      />
-    );
-  }, [
-    showPointTiles,
-    tileUrlTemplate,
-    withObsTiles,
-    withPressableObsTiles
-  ] );
 
   useEffect( ( ) => {
     if ( typeof ( onZoomChange ) === "function" ) {
@@ -496,94 +353,39 @@ const Map = ( {
         zoomEnabled={zoomEnabled}
         zoomTapEnabled={zoomTapEnabled}
       >
-        <ObsUrlTile />
-        {( showLocationIndicator && ( !obscured ) ) && (
-          <>
-            <Circle
-              center={{
-                latitude: obsLatitude,
-                longitude: obsLongitude
-              }}
-              radius={positionalAccuracy}
-              strokeWidth={2}
-              strokeColor={theme.colors.inatGreen}
-              fillColor="rgba( 116, 172, 0, 0.2 )"
-            />
-            <Marker
-              coordinate={{
-                latitude: obsLatitude,
-                longitude: obsLongitude
-              }}
-            >
-              {locationIndicator()}
-            </Marker>
-          </>
-        )}
-        {( showLocationIndicator && obscured ) && (
-          <Polygon
-            coordinates={[
-              {
-                latitude: obscurationCell.minLat,
-                longitude: obscurationCell.minLng
-              },
-              {
-                latitude: obscurationCell.minLat,
-                longitude: obscurationCell.maxLng
-              },
-              {
-                latitude: obscurationCell.maxLat,
-                longitude: obscurationCell.maxLng
-              },
-              {
-                latitude: obscurationCell.maxLat,
-                longitude: obscurationCell.minLng
-              }
-            ]}
-            strokeWidth={2}
-            strokeColor={colors.inatGreen}
-            fillColor="rgba( 116, 172, 0, 0.2 )"
-          />
-        )}
-      </MapView>
-      { showCurrentLocationButton && (
-        <INatIconButton
-          icon="location-crosshairs"
-          className={classnames(
-            "absolute bottom-5 right-5 bg-white rounded-full",
-            currentLocationButtonClassName
-          )}
-          style={DROP_SHADOW}
-          accessibilityLabel={t( "Zoom-to-current-location" )}
-          onPress={( ) => {
-            setZoomToUserLocationRequested( true );
-            setShowsUserLocation( true );
-            setPermissionRequested( true );
-          }}
+        <ObsUrlTile
+          params={params}
+          showPointTiles={currentZoom > 13}
+          withObsTiles={withObsTiles}
+          withPressableObsTiles={withPressableObsTiles}
         />
-      )}
-      {showSwitchMapTypeButton
-        && (
-          <INatIconButton
-            icon="map-layers"
-            className={classnames(
-              "absolute bottom-5 left-5 bg-white rounded-full",
-              switchMapTypeButtonClassName
-            )}
-            style={DROP_SHADOW}
-            accessibilityLabel={t( "Toggle-map-type" )}
-            accessibilityRole="button"
-            accessibilityState={
-              currentMapType === "standard"
-                ? t( "Standard--map-type" )
-                : t( "Satellite--map-type" )
-            }
-            onPress={( ) => {
-              changeMapType( currentMapType === "standard"
-                ? "hybrid"
-                : "standard" );
-            }}
-          />
-        )}
+        <LocationIndicator
+          obsLatitude={obsLatitude}
+          obsLongitude={obsLongitude}
+          positionalAccuracy={positionalAccuracy}
+          showLocationIndicator={showLocationIndicator && !obscured}
+        />
+        <ObscuredLocationIndicator
+          obscurationCell={obscurationCell}
+          showLocationIndicator={showLocationIndicator && obscured}
+        />
+      </MapView>
+      <CurrentLocationButton
+        showCurrentLocationButton={showCurrentLocationButton}
+        currentLocationButtonClassName={currentLocationButtonClassName}
+        handlePress={( ) => {
+          setZoomToUserLocationRequested( true );
+          setShowsUserLocation( true );
+          setPermissionRequested( true );
+        }}
+      />
+      <SwitchMapTypeButton
+        currentMapType={currentMapType}
+        mapType={mapType}
+        setCurrentMapType={setCurrentMapType}
+        showSwitchMapTypeButton={showSwitchMapTypeButton}
+        switchMapTypeButtonClassName={switchMapTypeButtonClassName}
+      />
       <LocationPermissionGate
         permissionNeeded={permissionRequested}
         onPermissionGranted={onPermissionGranted}
