+++ src/components/SharedComponents/index.js
@@ -15,7 +15,6 @@ export { default as Checkbox } from "./Checkbox";
 export { default as ConfidenceInterval } from "./ConfidenceInterval";
 export { default as DateDisplay } from "./DateDisplay";
 export { default as DateTimePicker } from "./DateTimePicker";
-export { default as DetailsMap } from "./DetailsMap";
 export { default as DisplayTaxon } from "./DisplayTaxon";
 export { default as DisplayTaxonName } from "./DisplayTaxonName";
 export { default as Divider } from "./Divider/Divider";
@@ -28,7 +27,8 @@ export { default as InlineUser } from "./InlineUser/InlineUser";
 export { default as InputField } from "./InputField";
 export { default as KebabMenu } from "./KebabMenu";
 export { default as LocationPermissionGate } from "./LocationPermissionGate";
-export { default as Map } from "./Map";
+export { default as DetailsMap } from "./Map/DetailsMap";
+export { default as Map } from "./Map/Map";
 export { default as MediaNavButtons } from "./MediaNavButtons";
 export { default as Modal } from "./Modal";
 export { default as Mortal } from "./Mortal";
